Add derive macro
