+++ Cargo.toml
@@ -52,7 +52,6 @@ opentelemetry-system-metrics = { version = "0.2.0", optional = true }
 tailcall-http-cache = { path = "tailcall-http-cache", optional = true }
 tailcall-version = { path = "./tailcall-version", optional = true }
 
-
 # dependencies safe for wasm:
 
 rustls-pemfile = { version = "1.0.4" }
@@ -161,6 +160,7 @@ Inflector = "0.11.4"
 path-clean = "=1.0.1"
 pathdiff = "0.2.1"
 num = "0.4.3"
+tailcall-typedefs-common = { path = "./tailcall-typedefs-common" }
 
 [dev-dependencies]
 tailcall-prettier = { path = "tailcall-prettier" }
@@ -213,6 +213,7 @@ default = ["cli", "js"]
 [workspace]
 members = [
     ".",
+    "tailcall-typedefs-common",
     "tailcall-typedefs",
     "tailcall-aws-lambda",
     "tailcall-cloudflare",
