+++ src/core/config/mod.rs
@@ -14,7 +14,7 @@ mod config_module;
 pub mod cors;
 mod from_document;
 pub mod group_by;
-mod headers;
+pub mod headers;
 mod into_document;
 mod key_values;
 mod link;
