+++ src/core/config/telemetry.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
+use tailcall_macros::DocumentDefinition;
 
 use super::KeyValue;
 use crate::core::config::{Apollo, ConfigReaderContext};
@@ -19,7 +20,18 @@ mod defaults {
 }
 
 /// Output the opentelemetry data to the stdout. Mostly used for debug purposes
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema, MergeRight)]
+#[derive(
+    Debug,
+    Clone,
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    schemars::JsonSchema,
+    MergeRight,
+    DocumentDefinition,
+)]
+#[doc_type("Input")]
 #[serde(rename_all = "camelCase")]
 pub struct StdoutExporter {
     /// Output to stdout in pretty human-readable format
@@ -28,7 +40,18 @@ pub struct StdoutExporter {
 }
 
 /// Output the opentelemetry data to otlp collector
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema, MergeRight)]
+#[derive(
+    Debug,
+    Clone,
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    schemars::JsonSchema,
+    MergeRight,
+    DocumentDefinition,
+)]
+#[doc_type("Input")]
 #[serde(rename_all = "camelCase")]
 pub struct OtlpExporter {
     pub url: String,
@@ -37,7 +60,18 @@ pub struct OtlpExporter {
 }
 
 /// Output format for prometheus data
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema)]
+#[derive(
+    Debug,
+    Clone,
+    Default,
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    schemars::JsonSchema,
+    DocumentDefinition,
+)]
+#[doc_type("Input")]
 #[serde(rename_all = "camelCase")]
 pub enum PrometheusFormat {
     #[default]
@@ -46,7 +80,10 @@ pub enum PrometheusFormat {
 }
 
 /// Output the telemetry metrics data to prometheus server
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema)]
+#[derive(
+    Debug, Clone, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema, DocumentDefinition,
+)]
+#[doc_type("Input")]
 #[serde(rename_all = "camelCase")]
 pub struct PrometheusExporter {
     #[serde(
@@ -58,7 +95,18 @@ pub struct PrometheusExporter {
     pub format: PrometheusFormat,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema, MergeRight)]
+#[derive(
+    Debug,
+    Clone,
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    schemars::JsonSchema,
+    MergeRight,
+    DocumentDefinition,
+)]
+#[doc_type("Input")]
 #[serde(rename_all = "camelCase")]
 pub enum TelemetryExporter {
     Stdout(StdoutExporter),
@@ -67,7 +115,18 @@ pub enum TelemetryExporter {
     Apollo(Apollo),
 }
 
-#[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq, schemars::JsonSchema)]
+#[derive(
+    Debug,
+    Default,
+    Clone,
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    schemars::JsonSchema,
+    DocumentDefinition,
+)]
+#[doc_type("DirectiveWithInput")]
 #[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase")]
 /// The @telemetry directive facilitates seamless integration with
