+++ src/core/config/upstream.rs
@@ -2,6 +2,7 @@ use std::collections::BTreeSet;
 
 use derive_setters::Setters;
 use serde::{Deserialize, Serialize};
+use tailcall_macros::DocumentDefinition;
 
 use crate::core::is_default;
 use crate::core::macros::MergeRight;
@@ -10,8 +11,18 @@ use crate::core::merge_right::MergeRight;
 const DEFAULT_MAX_SIZE: usize = 100;
 
 #[derive(
-    Serialize, Deserialize, PartialEq, Eq, Clone, Debug, Setters, schemars::JsonSchema, MergeRight,
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    Setters,
+    schemars::JsonSchema,
+    MergeRight,
+    DocumentDefinition,
 )]
+#[doc_type("Input")]
 #[serde(rename_all = "camelCase", default)]
 pub struct Batch {
     pub delay: usize,
@@ -29,7 +40,18 @@ impl Default for Batch {
     }
 }
 
-#[derive(Serialize, Deserialize, PartialEq, Eq, Clone, Debug, schemars::JsonSchema, MergeRight)]
+#[derive(
+    Serialize,
+    Deserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    schemars::JsonSchema,
+    MergeRight,
+    DocumentDefinition,
+)]
+#[doc_type("Input")]
 pub struct Proxy {
     pub url: String,
 }
@@ -45,7 +67,9 @@ pub struct Proxy {
     Default,
     schemars::JsonSchema,
     MergeRight,
+    DocumentDefinition,
 )]
+#[doc_type("Directive")]
 #[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase", default)]
 /// The `upstream` directive allows you to control various aspects of the
