+++ src/core/scalar/bytes.rs
@@ -1,11 +1,13 @@
 use async_graphql_value::ConstValue;
 use schemars::schema::Schema;
 use schemars::{schema_for, JsonSchema};
+use tailcall_macros::DocumentDefinition;
 
 use crate::core::json::JsonLike;
 
 /// Represents list of bytes
-#[derive(JsonSchema, Default)]
+#[derive(JsonSchema, Default, DocumentDefinition)]
+#[doc_type("Scalar")]
 pub struct Bytes(pub String);
 
 impl super::Scalar for Bytes {
