+++ src/core/scalar/int16.rs
@@ -1,9 +1,11 @@
 use async_graphql_value::ConstValue;
 use schemars::schema::Schema;
 use schemars::{schema_for, JsonSchema};
+use tailcall_macros::DocumentDefinition;
 
 /// Represents signed integer type 16bit size
-#[derive(JsonSchema, Default)]
+#[derive(JsonSchema, Default, DocumentDefinition)]
+#[doc_type("Scalar")]
 pub struct Int16(pub i16);
 
 impl super::Scalar for Int16 {
