+++ src/core/scalar/json.rs
@@ -1,12 +1,14 @@
 use async_graphql_value::ConstValue;
 use schemars::schema::Schema;
 use schemars::{schema_for, JsonSchema};
+use tailcall_macros::DocumentDefinition;
 
 /// The JSON scalar type represents JSON values as specified by
 /// [ECMA-404](www.ecma-international.org/publications/files/ECMA-ST/
 /// ECMA-404.pdf).
-#[derive(JsonSchema, Default)]
+#[derive(JsonSchema, Default, DocumentDefinition)]
 #[allow(clippy::upper_case_acronyms)]
+#[doc_type("Scalar")]
 pub struct JSON {
     #[allow(dead_code)]
     #[serde(rename = "JSON")]
