+++ tailcall-typedefs-common/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tailcall-typedefs"
+name = "tailcall-typedefs-common"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,12 +7,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = "1.0.82"
 async-graphql-parser = "7.0.6"
-lazy_static = "1.4.0"
+async-graphql-value = "7.0.6"
 schemars = "0.8.17"
-serde = "1.0.200"
-serde_json = "1.0.116"
-tailcall = { path = "../" }
-tokio = { version = "1.37.0", features = ["full"] }
-tracing = "0.1.40"
